requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
pyyaml>=6.0
requests>=2.25.0
beautifulsoup4>=4.9.0